refractor moving parts etc
