+++ src/App.js
@@ -1,9 +1,11 @@
 import React, { Component } from 'react';
 import CreatePost from "./components/createPost"
+import Post from "./components/post"
 
 import './App.css';
 import 'font-awesome/css/font-awesome.min.css';
 import { thisExpression } from '@babel/types';
+import PostList from "./components/postList"
 
 class App extends Component {
   constructor(props) {
@@ -25,57 +27,57 @@ class App extends Component {
 
   }
 
-  onContentChange = (e) => {
-    const name = e.target.name
-    const content = e.target.value;
-    console.log(content)
-    this.setState({
-      [name]: content
-    })
-  }
+//   onContentChange = (e) => {
+//     const name = e.target.name
+//     const content = e.target.value;
+//     console.log(content)
+//     this.setState({
+//       [name]: content
+//     })
+//   }
 
-  postSubmit = (e) => {
-    e.preventDefault();
-    const posts = this.state.posts;
-    const newPost = {
-      author: this.state.author,
-      content: this.state.content,
-      title: this.state.title,
-      voteCount: 0
-    }
-    posts.push(newPost)
-    // taking the posts array defined in original set of state and pushing content defined by user into the array and setting that in state
-    this.setState({
-      posts: posts,
-      content: "",
-      title: ""
-    })
-  }
+//   postSubmit = (e) => {
+//     e.preventDefault();
+//     const posts = this.state.posts;
+//     const newPost = {
+//       author: this.state.author,
+//       content: this.state.content,
+//       title: this.state.title,
+//       voteCount: 0
+//     }
+//     posts.push(newPost)
+//     // taking the posts array defined in original set of state and pushing content defined by user into the array and setting that in state
+//     this.setState({
+//       posts: posts,
+//       content: "",
+//       title: ""
+//     })
+//   }
 
 
-  vote = (e, sentPost, operator) => {
-    e.preventDefault();
-    const posts = this.state.posts.filter(checkPost => sentPost.title !== checkPost.title);
+//   vote = (e, sentPost, operator) => {
+//     e.preventDefault();
+//     const posts = this.state.posts.filter(checkPost => sentPost.title !== checkPost.title);
 
-    switch (operator) {
-      case "plus":
-        sentPost.voteCount++
-        break;
-      case "minus":
-        sentPost.voteCount--
-        break;
-      default:
-        console.error("you fucked it up")
-    }
+//     switch (operator) {
+//       case "plus":
+//         sentPost.voteCount++
+//         break;
+//       case "minus":
+//         sentPost.voteCount--
+//         break;
+//       default:
+//         console.error("you fucked it up")
+//     }
 
-    posts.push(sentPost);
-    posts.sort((a,b) => b.voteCount - a.voteCount)
+//     posts.push(sentPost);
+//     posts.sort((a,b) => b.voteCount - a.voteCount)
 
-    this.setState({
-      posts: posts
-    })
+//     this.setState({
+//       posts: posts
+//     })
   
-}
+// }
 
 
 
@@ -111,30 +113,21 @@ class App extends Component {
     console.log(this.state.posts)
     return (
       <div className="App">
-        <h1>Reddit</h1>
-        <CreatePost
-          postSubmit={this.postSubmit}
-          onContentChange={this.onContentChange}
-          title={this.state.title}
-          content={this.state.content}
-        />
-        <br />
-            
-        {this.state.posts.map((post, key) =>
-          <div key={key} className="post-wrapper">
-            <h4>{post.title}</h4>
-            <p> {post.content}</p>
-            <p>{post.voteCount}</p>
-            {/* <i className="fa fa-angle-double-down" onClick={(e) => this.vote(e, post, "plus")}></i> */}
-            <button onClick={(e) => this.vote(e, post, "plus")}>VOTE UP!</button>
-            <button onClick={(e) => this.vote(e, post, "minus")}>FUCK this! </button>
-            <br></br>
+       
+        <h1>R E D D I T  YA'BISH</h1>
+        <Post/>
+        
                 
           </div>
-        )}
-      </div>
+        
+      
     );
+    
   }
+
+  
+
+
 }
 
 export default App;
